+++ src/tr1/game/requester.c
@@ -135,8 +135,8 @@ int32_t Requester_Display(REQUEST_INFO *req)
 
     if (req->line_offset) {
         if (!req->moreup) {
-            req->moreup =
-                Text_Create(req->x, line_one_off - req->line_height + 2, "[");
+            req->moreup = Text_Create(
+                req->x, line_one_off - req->line_height + 2, "\\{arrow up}");
             Text_SetScale(
                 req->moreup, TEXT_BASE_SCALE * 2 / 3, TEXT_BASE_SCALE * 2 / 3);
             Text_CentreH(req->moreup, 1);
@@ -149,7 +149,7 @@ int32_t Requester_Display(REQUEST_INFO *req)
 
     if (req->items_used > req->vis_lines + req->line_offset) {
         if (!req->moredown) {
-            req->moredown = Text_Create(req->x, edge_y - 12, "]");
+            req->moredown = Text_Create(req->x, edge_y - 12, "\\{arrow down}");
             Text_SetScale(
                 req->moredown, TEXT_BASE_SCALE * 2 / 3,
                 TEXT_BASE_SCALE * 2 / 3);