+++ src/tr1/game/text.c
@@ -6,54 +6,6 @@
 #include "global/vars.h"
 
 #define TEXT_BOX_OFFSET 2
-#define TRIANGLE_SYM 93
-#define CIRCLE_SYM 94
-#define X_SYM 95
-#define SQUARE_SYM 96
-#define L2_SYM 99
-#define RIGHT_ARROW_SYM 109
-#define DOWN_ARROW_SYM 106
-#define UP_ARROW_SYM 107
-#define LEFT_ARROW_SYM 108
-#define RIGHT_ARROW_SYM 109
-
-static CLOCK_TIMER m_FlashTimer = { 0 };
-
-static int8_t m_TextSpacing[110] = {
-    14 /*A*/,  11 /*B*/, 11 /*C*/, 11 /*D*/, 11 /*E*/, 11 /*F*/, 11 /*G*/,
-    13 /*H*/,  8 /*I*/,  11 /*J*/, 12 /*K*/, 11 /*L*/, 13 /*M*/, 13 /*N*/,
-    12 /*O*/,  11 /*P*/, 12 /*Q*/, 12 /*R*/, 11 /*S*/, 12 /*T*/, 13 /*U*/,
-    13 /*V*/,  13 /*W*/, 12 /*X*/, 12 /*Y*/, 11 /*Z*/, 9 /*a*/,  9 /*b*/,
-    9 /*c*/,   9 /*d*/,  9 /*e*/,  9 /*f*/,  9 /*g*/,  9 /*h*/,  5 /*i*/,
-    9 /*j*/,   9 /*k*/,  5 /*l*/,  12 /*m*/, 10 /*n*/, 9 /*o*/,  9 /*p*/,
-    9 /*q*/,   8 /*r*/,  9 /*s*/,  8 /*t*/,  9 /*u*/,  9 /*v*/,  11 /*w*/,
-    9 /*x*/,   9 /*y*/,  9 /*z*/,  12 /*0*/, 8 /*1*/,  10 /*2*/, 10 /*3*/,
-    10 /*4*/,  10 /*5*/, 10 /*6*/, 9 /*7*/,  10 /*8*/, 10 /*9*/, 5 /*.*/,
-    5 /*,*/,   5 /*!*/,  11 /*?*/, 9 /*"*/,  10 /*"*/, 8 /**/,   6 /*(*/,
-    6 /*)*/,   7 /*-*/,  7 /*=*/,  3 /*:*/,  11 /*%*/, 8 /*+*/,  13 /*(c)*/,
-    16 /*tm*/, 9 /*&*/,  4 /*'*/,  12,       12,       7,        5,
-    7,         7,        7,        7,        7,        7,        7,
-    7,         16,       16,       16,       16,       16,       16,
-    16,        16,       16,       12,       14,       8,        8,
-    8,         8,        8,        8,        8
-};
-
-static int8_t m_TextASCIIMap[95] = {
-    0 /* */,   64 /*!*/,  66 /*"*/,  78 /*#*/, 77 /*$*/, 74 /*%*/, 78 /*&*/,
-    79 /*'*/,  69 /*(*/,  70 /*)*/,  92 /***/, 72 /*+*/, 63 /*,*/, 71 /*-*/,
-    62 /*.*/,  68 /**/,   52 /*0*/,  53 /*1*/, 54 /*2*/, 55 /*3*/, 56 /*4*/,
-    57 /*5*/,  58 /*6*/,  59 /*7*/,  60 /*8*/, 61 /*9*/, 73 /*:*/, 73 /*;*/,
-    66 /*<*/,  74 /*=*/,  75 /*>*/,  65 /*?*/, 0 /**/,   0 /*A*/,  1 /*B*/,
-    2 /*C*/,   3 /*D*/,   4 /*E*/,   5 /*F*/,  6 /*G*/,  7 /*H*/,  8 /*I*/,
-    9 /*J*/,   10 /*K*/,  11 /*L*/,  12 /*M*/, 13 /*N*/, 14 /*O*/, 15 /*P*/,
-    16 /*Q*/,  17 /*R*/,  18 /*S*/,  19 /*T*/, 20 /*U*/, 21 /*V*/, 22 /*W*/,
-    23 /*X*/,  24 /*Y*/,  25 /*Z*/,  80 /*[*/, 76 /*\*/, 81 /*]*/, 97 /*^*/,
-    98 /*_*/,  77 /*`*/,  26 /*a*/,  27 /*b*/, 28 /*c*/, 29 /*d*/, 30 /*e*/,
-    31 /*f*/,  32 /*g*/,  33 /*h*/,  34 /*i*/, 35 /*j*/, 36 /*k*/, 37 /*l*/,
-    38 /*m*/,  39 /*n*/,  40 /*o*/,  41 /*p*/, 42 /*q*/, 43 /*r*/, 44 /*s*/,
-    45 /*t*/,  46 /*u*/,  47 /*v*/,  48 /*w*/, 49 /*x*/, 50 /*y*/, 51 /*z*/,
-    100 /*{*/, 101 /*|*/, 102 /*}*/, 67 /*~*/
-};
 
 static RGBA_8888 m_MenuColorMap[MC_NUMBER_OF] = {
     { 70, 30, 107, 230 }, // MC_PURPLE_C
@@ -78,7 +30,6 @@ typedef struct {
     int32_t h;
 } QUAD_INFO;
 
-static uint8_t M_MapLetterToSpriteNum(char letter);
 static void M_DrawTextBackground(
     UI_STYLE ui_style, int32_t sx, int32_t sy, int32_t w, int32_t h,
     TEXT_STYLE text_style);
@@ -155,35 +106,6 @@ static void M_DrawTextOutline(
     }
 }
 
-static uint8_t M_MapLetterToSpriteNum(char letter)
-{
-    if (letter >= 16) {
-        return m_TextASCIIMap[letter - 32];
-    } else if (letter == '\200') {
-        return LEFT_ARROW_SYM;
-    } else if (letter == '\201') {
-        return RIGHT_ARROW_SYM;
-    } else if (letter == '\202') {
-        return DOWN_ARROW_SYM;
-    } else if (letter == '\203') {
-        return UP_ARROW_SYM;
-    } else if (letter == '\204') {
-        return TRIANGLE_SYM;
-    } else if (letter == '\205') {
-        return CIRCLE_SYM;
-    } else if (letter == '\206') {
-        return X_SYM;
-    } else if (letter == '\207') {
-        return SQUARE_SYM;
-    } else if (letter == '\300') {
-        return L2_SYM;
-    } else if (letter >= 11) {
-        return letter + 91;
-    } else {
-        return letter + 81;
-    }
-}
-
 RGBA_8888 Text_GetMenuColor(MENU_COLOR color)
 {
     return m_MenuColorMap[color];
@@ -211,30 +133,26 @@ int32_t Text_GetWidth(const TEXTSTRING *const text)
     }
 
     int32_t width = 0;
-    for (const char *ptr = text->content; *ptr != '\0'; *ptr++) {
-        const char letter = *ptr;
-        if (letter == 0x7F || (letter > 10 && letter < 32)) {
-            continue;
-        }
-
-        if (letter == 32) {
-            width += text->word_spacing;
-            continue;
-        }
-
-        uint8_t sprite_num = M_MapLetterToSpriteNum(letter);
-        width += m_TextSpacing[sprite_num] + text->letter_spacing;
+    const GLYPH_INFO **glyph_ptr = text->glyphs;
+    if (text->glyphs == NULL) {
+        return 0;
     }
+    while (*glyph_ptr != NULL) {
+        if ((*glyph_ptr)->role == GLYPH_SPACE) {
+            width += text->word_spacing;
+        } else {
+            width += (*glyph_ptr)->width + text->letter_spacing;
+        }
+        glyph_ptr++;
+    }
+
     width -= text->letter_spacing;
-    width &= 0xFFFE;
     return width * text->scale.h / TEXT_BASE_SCALE;
 }
 
 void Text_DrawText(TEXTSTRING *text)
 {
-    int sx, sy, sh, sv;
-
-    if (text->flags.hide) {
+    if (text->flags.hide || text->glyphs == NULL) {
         return;
     }
 
@@ -247,7 +165,6 @@ void Text_DrawText(TEXTSTRING *text)
         }
     }
 
-    char *content = text->content;
     double x = text->pos.x;
     double y = text->pos.y;
     int32_t text_width = Text_GetWidth(text);
@@ -268,41 +185,41 @@ void Text_DrawText(TEXTSTRING *text)
     int32_t bypos =
         text->background.offset.y + y - TEXT_BOX_OFFSET * 2 - TEXT_HEIGHT;
 
-    int32_t start_x = x;
+    int32_t sx;
+    int32_t sy;
+    int32_t sh;
+    int32_t sv;
+    const int32_t start_x = x;
 
-    int32_t letter = '\0';
-    while (*content) {
-        letter = *content++;
-        if (letter > 15 && letter < 32) {
-            continue;
-        }
-
-        if (text->flags.multiline && letter == '\n') {
+    const GLYPH_INFO **glyph_ptr = text->glyphs;
+    while (*glyph_ptr != NULL) {
+        if (text->flags.multiline && (*glyph_ptr)->role == GLYPH_NEWLINE) {
             y += TEXT_HEIGHT_FIXED * text->scale.v / TEXT_BASE_SCALE;
             x = start_x;
+            glyph_ptr++;
             continue;
         }
-        if (letter == ' ') {
+        if ((*glyph_ptr)->role == GLYPH_SPACE) {
             x += text->word_spacing * text->scale.h / TEXT_BASE_SCALE;
+            glyph_ptr++;
             continue;
         }
 
-        const uint8_t sprite_num = M_MapLetterToSpriteNum(letter);
         sx = Screen_GetRenderScale(x, RSR_TEXT);
         sy = Screen_GetRenderScale(y, RSR_TEXT);
         sh = Screen_GetRenderScale(text->scale.h, RSR_TEXT);
         sv = Screen_GetRenderScale(text->scale.v, RSR_TEXT);
 
         Output_DrawScreenSprite2D(
-            sx, sy, 0, sh, sv, g_Objects[O_ALPHABET].mesh_idx + sprite_num,
-            16 << 8, 0, 0);
+            sx, sy, 0, sh, sv,
+            g_Objects[O_ALPHABET].mesh_idx + (*glyph_ptr)->mesh_idx, 16 << 8, 0,
+            0);
 
-        if (letter == '(' || letter == ')' || letter == '$' || letter == '~') {
-            continue;
+        if ((*glyph_ptr)->role != GLYPH_COMBINING) {
+            x += (text->letter_spacing + (*glyph_ptr)->width) * text->scale.h
+                / TEXT_BASE_SCALE;
         }
-
-        x += ((int32_t)text->letter_spacing + m_TextSpacing[sprite_num])
-            * text->scale.h / TEXT_BASE_SCALE;
+        glyph_ptr++;
     }
 
     int32_t bwidth = 0;