+++ src/tr2/game/ui/widgets/prompt.c
@@ -2,7 +2,7 @@
 
 const char *UI_Prompt_GetPromptChar(void)
 {
-    return "\x11";
+    return "\\{button left}";
 }
 
 int32_t UI_Prompt_GetCaretFlashRate(void)