+++ src/tr2/global/vars_decomp.h
@@ -97,8 +97,6 @@
 #define g_KeyholePosition (*(XYZ_32*)0x004660E0)
 #define g_PuzzleHoleBounds (*(int16_t(*)[])0x004660F0)
 #define g_PuzzleHolePosition (*(XYZ_32*)0x00466108)
-#define g_TextSpacing (*(int8_t(*)[80])0x00466290)
-#define g_TextASCIIMap (*(int8_t(*)[])0x004662E0)
 #define g_BGND_PaletteIndex (*(int32_t*)0x00466400) // = -1
 #define g_GameClassName (*(const char(*)[])0x00466448)
 #define g_GameWindowName (*(const char(*)[])0x00466468)